requirements.txt
discord.py
requests
Pillow
flask
replit